feat: add release workflow

+++ scripts/build.sh
@@ -1,11 +1,25 @@
 #!/usr/bin/env bash
 
+if [ -z "$GOOS" ]; then
+	echo "Error: GOOS environment variable is not set."
+	exit 1
+fi
+
+if [ -z "$GOARCH" ]; then
+	echo "Error: GOOS environment variable is not set."
+	exit 1
+fi
+
 set +e
 rm -rf output
 
 set -e
 mkdir -p output/atsa-notifier
-go build -o ./output/atsa-notifier ./cmd/atsa-notifier
+if [ "$GOOS" = "windows" ]; then
+	go build -o ./output/atsa-notifier.exe ./cmd/atsa-notifier
+else
+	go build -o ./output/atsa-notifier ./cmd/atsa-notifier
+fi
 cp -r ./web ./output/atsa-notifier
 cp -r ./conf ./output/atsa-notifier
 cp -r ./data ./output/atsa-notifier
